Excel analysis and PDF explaining project and analysis
The Excel file contains five sheets of exercises about the impact economic variables have on the Opioid Crisis as well as illustrating total and average costs of opioid deaths over 50 states and from 2000-2018. Lastly the PDF will file is a full synopsis of the process involved in getting these results as well as going into more depth about the excel analysis.